fixed errors in the programm

+++ taskaev.matvei/P3/formingString.cpp
@@ -1,9 +1,9 @@
 #include "formingString.hpp"
 
-char* taskaev::formingString(const char* string, size_t& size, size_t newSize)
+char* taskaev::formingString(const char* string, int& size, int newSize)
 {
   char* newString = new char[size + 10];
-  for (size_t i = 0; i < newSize; i++)
+  for (int i = 0; i < newSize; i++)
   {
 	newString[i] = string[i];
   }
